Smart-HousePrice-Prediction-System-main/backend/requirements.txt

fastapi
uvicorn
requests
python-dotenv
sqlalchemy
pydantic
groq
scrapingbee